+++ feature/condregister/maybe_c2n.go
@@ -1,0 +1,8 @@
+// Copyright (c) Tailscale Inc & AUTHORS
+// SPDX-License-Identifier: BSD-3-Clause
+
+//go:build !ts_omit_c2n
+
+package condregister
+
+import _ "tailscale.com/feature/c2n"
